+++ CHANGELOG.md
@@ -56,6 +56,7 @@ BUG FIXES:
     rather than at runtime.
   - guests/redhat: Set hostname to FQDN, per the documentation for RedHat.
     [GH-2792]
+  - hosts/bsd: Don't invoke shell for NFS sudo calls. [GH-2808]
   - providers/virtualbox: Enabling internal networks by just setting "true"
     works properly. [GH-2751]
   - provisioners/chef: When chowning folders, don't follow symlinks.